+++ test/Bootstrap.php
@@ -1,114 +1,15 @@
 <?php
 
-namespace Eye4web\ZfcUserPmtest;
-
-use Zend\Loader\AutoloaderFactory;
-use Zend\Mvc\Service\ServiceManagerConfig;
-use Zend\ServiceManager\ServiceManager;
-use Zend\Stdlib\ArrayUtils;
-use RuntimeException;
-
-error_reporting(E_ALL | E_STRICT);
 chdir(__DIR__);
 
-class Bootstrap
-{
-    protected static $serviceManager;
-
-    public static function init()
-    {
-        // Load the user-defined test configuration file, if it exists; otherwise, load
-        if (is_readable(__DIR__.'/TestConfig.php')) {
-            $testConfig = include __DIR__.'/TestConfig.php';
-        } else {
-            $testConfig = include __DIR__.'/TestConfig.php.dist';
-        }
-
-        $zf2ModulePaths = [dirname(dirname(__DIR__))];
-        if (($path = static::findParentPath('vendor'))) {
-            $zf2ModulePaths[] = $path;
-        }
-        if (($path = static::findParentPath('module')) !== $zf2ModulePaths[0]) {
-            $zf2ModulePaths[] = $path;
-        }
-
-        $zf2ModulePaths  = implode(PATH_SEPARATOR, $zf2ModulePaths).PATH_SEPARATOR;
-        $zf2ModulePaths .= getenv('ZF2_MODULES_TEST_PATHS') ?:
-            (defined('ZF2_MODULES_TEST_PATHS') ?
-            ZF2_MODULES_TEST_PATHS : '');
-
-        static::initAutoloader();
-
-        // use ModuleManager to load this module and it's dependencies
-        $baseConfig = [
-            'module_listener_options' => [
-                'module_paths' => explode(PATH_SEPARATOR, $zf2ModulePaths),
-            ],
-        ];
-
-        $config = ArrayUtils::merge($baseConfig, $testConfig);
-
-        $serviceManager = new ServiceManager(new ServiceManagerConfig());
-        $serviceManager->setService('ApplicationConfig', $config);
-        $serviceManager->get('ModuleManager')->loadModules();
-        static::$serviceManager = $serviceManager;
-    }
-
-    public static function getServiceManager()
-    {
-        return static::$serviceManager;
-    }
-
-    protected static function initAutoloader()
-    {
-        $vendorPath = static::findParentPath('vendor');
-
-        if (is_readable($vendorPath.'/autoload.php')) {
-            $loader = include $vendorPath.'/autoload.php';
-        }
-
-        $zf2Path = getenv('ZF2_PATH') ?:
-            (defined('ZF2_PATH') ?
-            ZF2_PATH : (is_dir($vendorPath.'/ZF2/library') ?
-            $vendorPath.'/ZF2/library' : false));
-
-        if (!$zf2Path) {
-            throw new RuntimeException(
-                'Unable to load ZF2. Run `php composer.phar install` or define a ZF2_PATH environment variable.'
-            );
-        }
-
-        if (isset($loader)) {
-            $loader->add('Zend', $zf2Path.'/Zend');
-        } else {
-            include $zf2Path.'/Zend/Loader/AutoloaderFactory.php';
-            AutoloaderFactory::factory([
-                'Zend\Loader\StandardAutoloader' => [
-                    'autoregister_zf' => true,
-                    'namespaces' => [
-                        __NAMESPACE__ => __DIR__.'/'.__NAMESPACE__,
-                    ],
-                ],
-            ]);
-        }
-    }
-
-    protected static function findParentPath($path)
-    {
-        $dir = __DIR__;
-        $previousDir = '.';
-        while (!is_dir($dir.'/'.$path)) {
-            $dir = dirname($dir);
-
-            if ($previousDir === $dir) {
-                return false;
-            }
-
-            $previousDir = $dir;
-        }
-
-        return $dir.'/'.$path;
-    }
+$loader = null;
+if (file_exists('../../vendor/autoload.php')) { // for development / travis build
+    $loader = include '../../vendor/autoload.php';
+} elseif (file_exists('../../../vendor/autoload.php')) { // for usage in ZendSkeletonApplication
+    $loader = include '../../../vendor/autoload.php';
+} else {
+    throw new RuntimeException('vendor/autoload.php could not be found. Did you run `php composer.phar install`?');
 }
 
-Bootstrap::init();
+$loader->add('Eye4web\ZfcUserPmTest', __DIR__);
+include 'TestConfig.php';